+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 3.0.0-alpha.16 (2024-03-18)
 
 * (theimo1221) Improve behaviour of Strom-Stoss-Relais to prevent loops.
 * (theimo1221) Prepare Actuator "DayOn"-Setting